+++ PicoScopeICT/CustomClient/HWControl/scope_interface/PSDevice.h
@@ -22,9 +22,9 @@ public:
     void CollectOneWaveform();
     void GetTimebaseInfo(const int& timebase, int& samplingDurationInNanoseconds, int& samplingIntervalInNanoseconds);
 
-    int32_t currentWaveformChannelA[BUFFER_SIZE];
-    int32_t currentWaveformChannelB[BUFFER_SIZE];
-    int32_t currentTimeArray[BUFFER_SIZE];
+    float currentWaveformChannelA[BUFFER_SIZE];
+    float currentWaveformChannelB[BUFFER_SIZE];
+    float currentTimeArray[BUFFER_SIZE];
 
 private:
     int32_t timebase_{};
